+++ memberaudit/urls.py
@@ -207,9 +207,9 @@ urlpatterns = [
         name="corporation_compliance_report_data",
     ),
     path(
-        "skill_sets_report_data",
-        reports.skill_sets_report_data,
-        name="skill_sets_report_data",
+        "skill_sets_report_2_data",
+        reports.SkillSetReportDataTableView.as_view(),
+        name="skill_sets_report_2_data",
     ),
     # data export
     path("data-export/", data_export.data_export, name="data_export"),